refactor : 목표 수행 MVP 패턴 적용
- DoMyGoal -> Post로 네이밍 변경
- 패키지 이동
