+++ app/src/main/java/kr/co/company/capstone/post/contract/PostContract.kt
@@ -1,0 +1,4 @@
+package kr.co.company.capstone.post.contract
+
+class PostContract {
+}
