+++ app/src/main/java/kr/co/company/capstone/post/model/PostModel.kt
@@ -1,0 +1,4 @@
+package kr.co.company.capstone.post.model
+
+class PostModel {
+}
