+++ app/src/main/java/kr/co/company/capstone/post/presenter/PostPresenter.kt
@@ -1,0 +1,4 @@
+package kr.co.company.capstone.post.presenter
+
+class PostPresenter {
+}
